+++ DManager/Models/PreviewDebtModel.cs
@@ -3,6 +3,6 @@
     public class PreviewDebtModel
     {
         public string Name { set; get; }
-        public int DebtSum { set; get; }
+        public double DebtSum { set; get; }
     }
 }
